6주차 조회 API
